# production_backend/docs/openapi.generated.json
{
  "components": {
    "schemas": {
...
      "FeedingRecordCreate": {
        "properties": {
          "duration_seconds": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "feed_action": {
            "anyOf": [
              {
                "maxLength": 32,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feed Action"
          },
          "feed_time": {
            "format": "date-time",
            "title": "Feed Time",
            "type": "string"
          },
          "feed_type": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Feed Type",
            "type": "string"
          },
          "infant_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Infant Id"
          },
          "plan_task_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Task Id"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "volume_ml": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume Ml"
          }
        },
        "required": [
          "feed_time",
          "feed_type"
        ],
        "title": "FeedingRecordCreate",
        "type": "object"
      },
      "FeedingRecordListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FeedingRecordRead"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "FeedingRecordListResponse",
        "type": "object"
      },
      "FeedingRecordRead": {
        "properties": {
          "duration_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "feed_action": {
            "title": "Feed Action",
            "type": "string"
          },
          "feed_time": {
            "format": "date-time",
            "title": "Feed Time",
            "type": "string"
          },
          "feed_type": {
            "title": "Feed Type",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "infant_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Infant Id"
          },
          "owner_user_id": {
            "format": "uuid",
            "title": "Owner User Id",
            "type": "string"
          },
          "plan_task_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Task Id"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "volume_ml": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume Ml"
          }
        },
        "required": [
          "id",
          "owner_user_id",
          "feed_time",
          "feed_type",
          "feed_action",
          "title",
          "status"
        ],
        "title": "FeedingRecordRead",
        "type": "object"
      },
...
      "GrowthRecordCreate": {
        "properties": {
          "head_cm": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Head Cm"
          },
          "height_cm": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height Cm"
          },
          "infant_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Infant Id"
          },
          "measured_at": {
            "format": "date-time",
            "title": "Measured At",
            "type": "string"
          },
          "weight_kg": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Kg"
          }
        },
        "required": [
          "measured_at"
        ],
        "title": "GrowthRecordCreate",
        "type": "object"
      },
      "GrowthRecordListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/GrowthRecordRead"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "GrowthRecordListResponse",
        "type": "object"
      },
      "GrowthRecordRead": {
        "properties": {
          "head_cm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Head Cm"
          },
          "height_cm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height Cm"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "infant_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Infant Id"
          },
          "measured_at": {
            "format": "date-time",
            "title": "Measured At",
            "type": "string"
          },
          "owner_user_id": {
            "format": "uuid",
            "title": "Owner User Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "weight_kg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Kg"
          }
        },
        "required": [
          "id",
          "owner_user_id",
          "measured_at",
          "status"
        ],
        "title": "GrowthRecordRead",
        "type": "object"
      },
      "GrowthRecordUpdate": {
        "properties": {
          "head_cm": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Head Cm"
          },
          "height_cm": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height Cm"
          },
          "infant_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Infant Id"
          },
          "measured_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Measured At"
          },
          "weight_kg": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Kg"
          }
        },
        "title": "GrowthRecordUpdate",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MilkTrendDayRead": {
        "properties": {
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "measured_only": {
            "default": true,
            "title": "Measured Only",
            "type": "boolean"
          },
          "pumped_milk_volume_ml": {
            "title": "Pumped Milk Volume Ml",
            "type": "number"
          },
          "pumping_count": {
            "title": "Pumping Count",
            "type": "integer"
          }
        },
        "required": [
          "date",
          "pumped_milk_volume_ml",
          "pumping_count"
        ],
        "title": "MilkTrendDayRead",
        "type": "object"
      },
      "MilkTrendListResponse": {
        "properties": {
          "days": {
            "title": "Days",
            "type": "integer"
          },
          "include_today": {
            "title": "Include Today",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/MilkTrendDayRead"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items",
          "days",
          "include_today"
        ],
        "title": "MilkTrendListResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "MomCozy Production Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/records/feeding": {
      "get": {
        "operationId": "list_feedings_v1_records_feeding_get",
        "parameters": [
          {
            "in": "query",
            "name": "start_at",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start At"
            }
          },
          {
            "in": "query",
            "name": "end_at",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End At"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedingRecordListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Feedings",
        "tags": [
          "records"
        ],
        "x-momcozy-api-surface": "public_app_api",
        "x-momcozy-client": [
          "flutter"
        ],
        "x-momcozy-owner": "records",
        "x-momcozy-stability": "stable"
      },
      "post": {
        "operationId": "create_feeding_v1_records_feeding_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedingRecordCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedingRecordRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Feeding",
        "tags": [
          "records"
        ],
        "x-momcozy-api-surface": "public_app_api",
        "x-momcozy-client": [
          "flutter"
        ],
        "x-momcozy-owner": "records",
        "x-momcozy-stability": "stable"
      }
    },
    "/v1/records/feeding/{record_id}": {
      "delete": {
        "operationId": "delete_feeding_v1_records_feeding__record_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "record_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Record Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Feeding",
        "tags": [
          "records"
        ],
        "x-momcozy-api-surface": "public_app_api",
        "x-momcozy-client": [
          "flutter"
        ],
        "x-momcozy-owner": "records",
        "x-momcozy-stability": "stable"
      }
    },
    "/v1/records/growth": {
      "get": {
        "operationId": "list_growth_v1_records_growth_get",
        "parameters": [
          {
            "in": "query",
            "name": "infant_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Infant Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrowthRecordListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Growth",
        "tags": [
          "records"
        ],
        "x-momcozy-api-surface": "public_app_api",
        "x-momcozy-client": [
          "flutter"
        ],
        "x-momcozy-owner": "records",
        "x-momcozy-stability": "stable"
      },
      "post": {
        "operationId": "create_growth_v1_records_growth_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrowthRecordCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrowthRecordRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Growth",
        "tags": [
          "records"
        ],
        "x-momcozy-api-surface": "public_app_api",
        "x-momcozy-client": [
          "flutter"
        ],
        "x-momcozy-owner": "records",
        "x-momcozy-stability": "stable"
      }
    },
    "/v1/records/growth/{record_id}": {
      "delete": {
        "operationId": "delete_growth_v1_records_growth__record_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "record_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Record Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Growth",
        "tags": [
          "records"
        ],
        "x-momcozy-api-surface": "public_app_api",
        "x-momcozy-client": [
          "flutter"
        ],
        "x-momcozy-owner": "records",
        "x-momcozy-stability": "stable"
      },
      "patch": {
        "operationId": "update_growth_v1_records_growth__record_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "record_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Record Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrowthRecordUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrowthRecordRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Growth",
        "tags": [
          "records"
        ],
        "x-momcozy-api-surface": "public_app_api",
        "x-momcozy-client": [
          "flutter"
        ],
        "x-momcozy-owner": "records",
        "x-momcozy-stability": "stable"
      }
    },
    "/v1/records/milk-trends": {
      "get": {
        "operationId": "get_milk_trends_v1_records_milk_trends_get",
        "parameters": [
          {
            "in": "query",
            "name": "start_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 90,
              "minimum": 1,
              "title": "Days",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "include_today",
            "required": false,
            "schema": {
              "default": true,
              "title": "Include Today",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MilkTrendListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Milk Trends",
        "tags": [
          "records"
        ],
        "x-momcozy-api-surface": "public_app_api",
        "x-momcozy-client": [
          "flutter"
        ],
        "x-momcozy-owner": "records",
        "x-momcozy-stability": "stable"
      }
    },
...
  }
}